refactor: remove second fragment
