--- app/src/main/kotlin/xyz/lbres/androidapptemplate/ui/second/SecondViewModel.kt
@@ -1,10 +1,0 @@
-package xyz.lbres.androidapptemplate.ui.second
-
-import androidx.lifecycle.ViewModel
-
-/**
- * ViewModel for second fragment
- */
-class SecondViewModel : ViewModel() {
-    var visitedCount: Int = 0
-}
