+++ app/src/main/res/layout/fragment_home.xml
@@ -7,12 +7,14 @@
     android:background="?backgroundColor">
 
     <TextView
-        android:id="@+id/navigateText"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="@string/click_navigate"
+        android:layout_marginTop="80dp"
+        android:paddingHorizontal="20dp"
+        android:text="@string/welcome_message"
         android:textColor="?colorOnBackground"
-        app:layout_constraintBottom_toBottomOf="parent"
+        android:textSize="24sp"
+        android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
