+++ app/src/main/res/navigation/nav_graph.xml
@@ -1,31 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <navigation xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    app:startDestination="@id/firstFragment"
+    app:startDestination="@id/homeFragment"
     android:id="@+id/nav_graph">
 
     <fragment
-        android:id="@+id/firstFragment"
-        android:name="xyz.lbres.androidapptemplate.ui.first.FirstFragment"
-        android:label="FirstFragment">
+        android:id="@+id/homeFragment"
+        android:name="xyz.lbres.androidapptemplate.ui.home.HomeFragment"
+        android:label="HomeFragment">
 
         <action
-            android:id="@+id/navigateFirstToSecond"
-            app:destination="@id/secondFragment" />
-
-        <action
-            android:id="@+id/navigateFirstToAttributions"
+            android:id="@+id/navigateHomeToAttributions"
             app:destination="@id/attributionsFragment" />
-    </fragment>
-
-    <fragment
-        android:id="@+id/secondFragment"
-        android:name="xyz.lbres.androidapptemplate.ui.second.SecondFragment"
-        android:label="SecondFragment">
-
-        <action
-            android:id="@+id/navigateSecondToFirst"
-            app:destination="@id/firstFragment" />
     </fragment>
 
     <fragment
